sse functions modified to handle double precision instead single precision

+++ include/sse.hpp
@@ -1,6 +1,6 @@
 #include <immintrin.h>
 #include<stdio.h>
-float* sse_add(float*, float*, float*, int size);
-float* sse_sub(float*, float*, float*, int size);
-float* sse_mul(float*, float*, float*, int size);
+double* sse_add(double*, double*, double*, int size);
+double* sse_sub(double*, double*, double*, int size);
+double* sse_mul(double*, double*, double*, int size);
